add CRC to snapshot file format and validate on load
additionally while imlementing this it was discovered
that the index Config struct contains items which
have state, this is undesirable as the Config should
be easily shareable.  to address this, deletion policy
and directory have been replaced with functions that
create the appropriate implementation, and these are
held directly by the reader/writer outside of the
config.

fixes #2