quick fix sort issue (#87)
* quick fix to sort issue

It has been observed that in some cases the computed sort value
for a DocumentMatch becomes corrupted.  The problem has been
traced back to the doc values uncompressed slices, but for now
a quick fix is proposed to copy the bytes associate with the
sort key, ensuring that no other doc values operations can
corrupt them.